Create scope procedures and type check for returns
Adds the concepts of a scope procedure, which is the procedure that created
a given scope. Adds functions for retrieving and setting that scope procedure.

Adds type checking to the return statement, which must consult the scope
procedure, to make sure the return statement is the same type as the expression
type