Add fix for binary/logical opreator type checking
I suspect this is going to be a temporary fix. The problem here is that
for codegen I probably need to know what the destination type is going to be
in order to know how I should treat the '&', '|', and 'not' operators, because
they have different type rules for what the resulting type is.